feat: 피드 로그인관련 기능 추가

+++ BE/src/main/java/com/ssafy/foodthink/feed/service/FeedServiceImpl.java
@@ -1,6 +1,5 @@
 package com.ssafy.foodthink.feed.service;
 
-import com.ssafy.foodthink.another.*;
 import com.ssafy.foodthink.feed.dto.FeedCommentRequestDto;
 import com.ssafy.foodthink.feed.dto.FeedCommentResponseDto;
 import com.ssafy.foodthink.feed.dto.FeedRequestDto;
@@ -13,16 +12,20 @@ import com.ssafy.foodthink.feed.repository.FeedCommentRepository;
 import com.ssafy.foodthink.feed.repository.FeedImageRepository;
 import com.ssafy.foodthink.feed.repository.FeedLikeRepository;
 import com.ssafy.foodthink.feed.repository.FeedRepository;
+import com.ssafy.foodthink.feed.tmp.RecipeRepository;
+import com.ssafy.foodthink.global.S3Service;
+import com.ssafy.foodthink.recipes.entity.RecipeEntity;
+import com.ssafy.foodthink.user.entity.UserEntity;
+import com.ssafy.foodthink.user.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -30,24 +33,23 @@ import java.util.stream.Collectors;
 public class FeedServiceImpl implements FeedService{
 
     private final FeedRepository feedRepository;
-    private final UsersRepository usersRepository;
-    private final UserRecipeRepository userRecipeRepository;
-    private final CrawlingRecipeRepository crawlingRecipeRepository;
+    private final UserRepository userRepository;
+    private final RecipeRepository recipeRepository;
     private final S3Service s3Service;
     private final FeedImageRepository feedImageRepository;
     private final FeedLikeRepository feedLikeRepository;
     private final FeedCommentRepository feedCommentRepository;
 
-    public FeedServiceImpl(FeedRepository feedRepository, UsersRepository usersRepository, UserRecipeRepository userRecipeRepository, CrawlingRecipeRepository crawlingRecipeRepository, S3Service s3Service, FeedImageRepository feedImageRepository, FeedLikeRepository feedLikeRepository, FeedCommentRepository feedCommentRepository) {
+    public FeedServiceImpl(FeedRepository feedRepository, UserRepository userRepository, RecipeRepository recipeRepository, S3Service s3Service, FeedImageRepository feedImageRepository, FeedLikeRepository feedLikeRepository, FeedCommentRepository feedCommentRepository) {
         this.feedRepository = feedRepository;
-        this.usersRepository = usersRepository;
-        this.userRecipeRepository = userRecipeRepository;
-        this.crawlingRecipeRepository = crawlingRecipeRepository;
+        this.userRepository = userRepository;
+        this.recipeRepository = recipeRepository;
         this.s3Service = s3Service;
         this.feedImageRepository = feedImageRepository;
         this.feedLikeRepository = feedLikeRepository;
         this.feedCommentRepository = feedCommentRepository;
     }
+
 
     @Override
     @Transactional
@@ -56,27 +58,20 @@ public class FeedServiceImpl implements FeedService{
 //        log.info(feedDto);
         log.info(feedRequestDto.toString());
         // Optional에서 값 추출 (없으면 예외 던짐)
-        UsersEntity user = usersRepository.findUsersByUserId(feedRequestDto.getUserId())
-                .orElseThrow(() -> new IllegalArgumentException("User not found"));
-        // userRecipeId가 주어졌다면, 해당 UserRecipe를 찾고, 그렇지 않으면 null로 설정
-        UserRecipe userRecipe = feedRequestDto.getUserRecipeId() != null
-                ? userRecipeRepository.findById(feedRequestDto.getUserRecipeId())
-                .orElseThrow(() -> new IllegalArgumentException("UserRecipe not found"))
-                : null;
+        UserEntity user = userRepository.findByUserId(feedRequestDto.getUserId())
+                .orElseThrow(() -> new IllegalArgumentException("해당 사용자가 없습니다."));
 
-        // crawlingRecipeId가 주어졌다면, 해당 CrawlingRecipe를 찾고, 그렇지 않으면 null로 설정
-        CrawlingRecipe crawlingRecipe = feedRequestDto.getCrawlingRecipeId() != null
-                ? crawlingRecipeRepository.findById(feedRequestDto.getCrawlingRecipeId())
-                .orElseThrow(() -> new IllegalArgumentException("CrawlingRecipe not found"))
+        RecipeEntity recipe = feedRequestDto.getRecipeId() != null
+                ? recipeRepository.findById(feedRequestDto.getRecipeId())
+                .orElseThrow(() -> new IllegalArgumentException("해당 레시피가 존재하지 않습니다."))
                 : null;
 
 
         FeedEntity feedEntity = FeedEntity.builder()
                 .foodName(feedRequestDto.getFoodName())
                 .content(feedRequestDto.getContent())
-                .usersEntity(user)
-                .userRecipe(userRecipe)
-                .crawlingRecipe(crawlingRecipe)
+                .userEntity(user)
+                .recipeEntity(recipe)
                 .build();
 
         FeedEntity saveFeed = feedRepository.save(feedEntity);
@@ -115,10 +110,18 @@ public class FeedServiceImpl implements FeedService{
     }
 
     @Override
+    public FeedResponseDto readFeedById(Long id, Long userId) {
+        FeedResponseDto feedResponseDto = readFeedById(id);
+        boolean check = feedLikeRepository.existsByFeedEntity_IdAndUserEntity_userId(id, userId);
+        feedResponseDto.setLike(check);
+        return feedResponseDto;
+    }
+
+    @Override
     public List<FeedResponseDto> readFeedsByUserId(Long userId) {
         List<FeedResponseDto> feedResponseDtos = new ArrayList<>();
 
-        List<FeedEntity> feedEntities = feedRepository.findAllByUsersEntity_userIdOrderByWriteTime(userId);
+        List<FeedEntity> feedEntities = feedRepository.findAllByUserEntity_userIdOrderByWriteTime(userId);
 
         for (FeedEntity feedEntity : feedEntities) {
             List<String> imageUrls = readImageUrlsByFeedId(feedEntity.getId());
@@ -132,7 +135,7 @@ public class FeedServiceImpl implements FeedService{
     @Override
     public void createFeedLikeByFeedId(Long feedId, Long userId) {
         //중복 여부 확인
-        boolean exists = feedLikeRepository.existsByFeedEntity_IdAndUsersEntity_userId(feedId, userId);
+        boolean exists = feedLikeRepository.existsByFeedEntity_IdAndUserEntity_userId(feedId, userId);
         if(exists) {
             throw  new IllegalArgumentException(" 이미 좋아요를 누른 상태입니다.");
         }
@@ -140,13 +143,13 @@ public class FeedServiceImpl implements FeedService{
         //엔티티 조회
         FeedEntity feedEntity = feedRepository.findById(feedId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 피드를 찾을 수 없습니다. ID: " + feedId));
-        UsersEntity usersEntity = usersRepository.findUsersByUserId(userId)
+        UserEntity UserEntity = userRepository.findByUserId(userId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 유저를 찾을 수 없습니다. ID: " + userId));
 
         //좋아요 저장
         FeedLikeEntity feedLikeEntity = FeedLikeEntity.builder()
                 .feedEntity(feedEntity)
-                .usersEntity(usersEntity)
+                .userEntity(UserEntity)
                 .build();
 
         feedLikeRepository.save(feedLikeEntity);
@@ -154,7 +157,16 @@ public class FeedServiceImpl implements FeedService{
 
     @Override
     public void deleteFeedLikeByFeedId(Long feedId, Long userId) {
-        FeedLikeEntity feedLikeEntity = feedLikeRepository.findByFeedEntity_IdAndUsersEntity_userId(feedId, userId);
+        UserEntity UserEntity = userRepository.findByUserId(userId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 유저를 찾을 수 없습니다. ID: " + userId));
+
+        //좋아요 기록이 있는지 확인
+        boolean exists = feedLikeRepository.existsByFeedEntity_IdAndUserEntity_userId(feedId, userId);
+        if(!exists) {
+            throw  new IllegalArgumentException("이미 삭제된 피드 좋아요입니다.");
+        }
+
+        FeedLikeEntity feedLikeEntity = feedLikeRepository.findByFeedEntity_IdAndUserEntity_userId(feedId, userId);
         feedLikeRepository.delete(feedLikeEntity);
     }
 
@@ -163,13 +175,13 @@ public class FeedServiceImpl implements FeedService{
         //엔티티 조회
         FeedEntity feedEntity = feedRepository.findById(feedId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 피드를 찾을 수 없습니다. ID: " + feedId));
-        UsersEntity usersEntity = usersRepository.findUsersByUserId(feedCommentRequestDto.getUserId())
+        UserEntity UserEntity = userRepository.findByUserId(feedCommentRequestDto.getUserId())
                 .orElseThrow(() -> new IllegalArgumentException("해당 유저를 찾을 수 없습니다. ID: " + feedCommentRequestDto.getUserId()));
 
         //댓글 저장
         FeedCommentEntity feedCommentEntity = FeedCommentEntity.builder()
                 .content(feedCommentRequestDto.getContent())
-                .usersEntity(usersEntity)
+                .userEntity(UserEntity)
                 .feedEntity(feedEntity)
                 .build();
         feedCommentRepository.save(feedCommentEntity);
@@ -179,6 +191,7 @@ public class FeedServiceImpl implements FeedService{
     public void deleteFeedCommentByFeedCommentId(Long feedCommentId) {
         FeedCommentEntity feedCommentEntity = feedCommentRepository.findById(feedCommentId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 피드 댓글을 찾을 수 없습니다. ID: " + feedCommentId));
+
         feedCommentRepository.delete(feedCommentEntity);
     }
 
@@ -189,7 +202,7 @@ public class FeedServiceImpl implements FeedService{
                 .orElseThrow(() -> new IllegalArgumentException("해당 피드 댓글을 찾을 수 없습니다. ID: " + feedCommentId));
 
         //피드 댓글 작성자만 수정 가능
-        if(!Objects.equals(feedCommentEntity.getUsersEntity().getUserId(), feedCommentRequestDto.getUserId())){
+        if(!Objects.equals(feedCommentEntity.getUserEntity().getUserId(), feedCommentRequestDto.getUserId())){
             throw new RuntimeException("댓글을 작성한 사용자와 다릅니다.");
         }
 
@@ -209,17 +222,17 @@ public class FeedServiceImpl implements FeedService{
 
 
     @Override
-    public List<FeedResponseDto> readFeedsByUserIdAndLogIn(Long searchUserId, Long logInUserId) {
+    public List<FeedResponseDto> readFeedsByUserId(Long searchUserId, Long logInUserId) {
         List<FeedResponseDto> feedResponseDtos = new ArrayList<>();
 
-        List<FeedEntity> feedEntities = feedRepository.findAllByUsersEntity_userIdOrderByWriteTime(searchUserId);
+        List<FeedEntity> feedEntities = feedRepository.findAllByUserEntity_userIdOrderByWriteTime(searchUserId);
 
         for (FeedEntity feedEntity : feedEntities) {
             List<String> imageUrls = readImageUrlsByFeedId(feedEntity.getId());
 
             FeedResponseDto feedResponseDto = createFeedResponseDtoByBuilder(feedEntity, imageUrls);
             if(logInUserId != null) {
-                feedResponseDto.setLike(feedLikeRepository.existsByFeedEntity_IdAndUsersEntity_userId(feedResponseDto.getId(), logInUserId));
+                feedResponseDto.setLike(feedLikeRepository.existsByFeedEntity_IdAndUserEntity_userId(feedResponseDto.getId(), logInUserId));
             }
             feedResponseDtos.add(feedResponseDto);
         }
@@ -228,9 +241,15 @@ public class FeedServiceImpl implements FeedService{
     }
 
     @Override
-    public void deleteFeedByFeedId(Long feedId) {
+    public void deleteFeedByFeedId(Long feedId, Long userId) {
         FeedEntity feedEntity = feedRepository.findById(feedId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 피드가 존재하지 않습니다."));
+
+        // 피드 작성자와 요청한 사용자가 동일한지 확인
+        if (!feedEntity.getUserEntity().getUserId().equals(userId)) {
+            throw new AccessDeniedException("본인 피드만 삭제할 수 있습니다.");
+        }
+
         feedRepository.delete(feedEntity);
     }
 
@@ -251,10 +270,9 @@ public class FeedServiceImpl implements FeedService{
                 .foodName(feedEntity.getFoodName())
                 .content(feedEntity.getContent())
                 .writeTime(feedEntity.getWriteTime())
-                .userId(feedEntity.getUsersEntity().getUserId())
-                .username(feedEntity.getUsersEntity().getUsername())
-                .userRecipeId(feedEntity.getUserRecipe() != null ? feedEntity.getUserRecipe().getRecipeId() : null)
-                .crawlingRecipeId(feedEntity.getCrawlingRecipe() != null ? feedEntity.getCrawlingRecipe().getRecipeId() : null)
+                .userId(feedEntity.getUserEntity().getUserId())
+                .username(feedEntity.getUserEntity().getNickname())
+                .recipeId(feedEntity.getRecipeEntity() != null ? feedEntity.getRecipeEntity().getRecipeId() : null)
                 .images(imageUrls)
                 .build();
     }
